5-assistant/requirements_assistant.txt
openai>=1.40.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
numpy>=1.26.0
python-dotenv>=1.0.1
